First Commit
This is a blockchain simulator, using PoW consensus mechanism.
Miners are in as threads, mining all at the same time to get the right length of difficulty for the generated hash.
Clients are also runnables, they send messages randomly during the generation of the hash for the block, once the block is generated the clients are timed out.